+++ src/BindingMap.ts
@@ -1,0 +1,5 @@
+import { BindingMapInterface } from './interfaces/BindingMapInterface'
+
+export class BindingMap extends Map implements BindingMapInterface {
+
+}