+++ src/interfaces/BindingMapInterface.ts
@@ -1,0 +1,5 @@
+import { BindingInterface } from './bindings/BindingInterface'
+
+export interface BindingMapInterface extends Map<any, BindingInterface> {
+
+}