+++ src/interfaces/ContainerInterface.ts
@@ -1,4 +1,4 @@
-import { BindingInterface } from './bindings/BindingInterface'
+import { BindingMapInterface } from './BindingMapInterface'
 import { Provider } from '../types/Provider'
 import { Resolver } from '../types/Resolver'
 
@@ -6,41 +6,41 @@ export interface ContainerInterface {
   /**
    * Bind a single instance or value into the container under the provided key.
    *
-   * @param  {string} key
-   * @param  {any}    value
+   * @param  {any} key
+   * @param  {any} value
    * @return {this}
    */
-  instance (key: string, value: any): this
+  instance (key: any, value: any): this
 
   /**
    * Bind a resolver function into the container under the provided key. The
    * resolver will be run once and the resulting value will be returned for all
    * subsequent resolutions.
    *
-   * @param  {string}   key
+   * @param  {any}      key
    * @param  {Resolver} resolver
    * @return {this}
    */
-  singleton (key: string, resolver: Resolver): this
+  singleton (key: any, resolver: Resolver): this
 
   /**
    * Bind a resolver function into the container under the provided key. The
    * resolver will be run each time the key is resolved resulting in new
    * instances each resolution.
    *
-   * @param  {string}   key
+   * @param  {any}      key
    * @param  {Resolver} resolver
    * @return {this}
    */
-  binding (key: string, resolver: Resolver): this
+  binding (key: any, resolver: Resolver): this
 
   /**
    * Resolve a value from the container by its key.
    *
-   * @param  {string} key
+   * @param  {any} key
    * @return {any}
    */
-  make (key: string): any
+  make (key: any): any
 
   /**
    * Add a service provider into the container to register one or many bindings
@@ -64,9 +64,9 @@ export interface ContainerInterface {
   /**
    * Retrieve the value of the bindings property.
    *
-   * @return {{ [key: string]: BindingInterface }}
+   * @return {BindingMapInterface}
    */
-  getBindings (): { [key: string]: BindingInterface }
+  getBindings (): BindingMapInterface
 
   /**
    * Retrieve the value of the providers property.