+++ CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- `BindingMap` class with `BindingMapInterface`
+
+### Changed
+- Replaced generic object with `BindingMap` for holding bindings in `Container`
 
 ## [0.2.1]
 ### Changed